Move parseDate/getDate to src/core
